Finalizando ResponseEnt

+++ Contas/src/main/java/com/gerenciador/contas/service/CidadeService.java
@@ -3,7 +3,10 @@ package com.gerenciador.contas.service;
 import com.gerenciador.contas.model.CidadeModel;
 import com.gerenciador.contas.repository.CidadeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -31,7 +34,12 @@ public class CidadeService {
     }
 
     public List<CidadeModel> deletar(Long id){
-        cidadeReposirory.deleteById(id);
-        return buscarTodas();
+        if (cidadeReposirory.findById(id).isPresent()) {
+            cidadeReposirory.deleteById(id);
+            throw new HttpClientErrorException(HttpStatus.NO_CONTENT);
+        }else if (cidadeReposirory.findAll().isEmpty()){
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
+        }else
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
     }
 }
