+++ Contas/src/main/java/com/gerenciador/contas/service/EnderecoService.java
@@ -4,7 +4,10 @@ package com.gerenciador.contas.service;
 import com.gerenciador.contas.model.EnderecoModel;
 import com.gerenciador.contas.repository.EnderecoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -33,6 +36,12 @@ public class EnderecoService {
     }
 
     public void deletar(Long codigo){
-        enderecoRepository.deleteById(codigo);
+        if (enderecoRepository.findById(codigo).isPresent()) {
+            enderecoRepository.deleteById(codigo);
+            throw new HttpClientErrorException(HttpStatus.NO_CONTENT);
+        }else if (enderecoRepository.findAll().isEmpty()){
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
+        }else
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
     }
 }
