+++ Contas/src/main/java/com/gerenciador/contas/service/EstadoService.java
@@ -4,7 +4,10 @@ package com.gerenciador.contas.service;
 import com.gerenciador.contas.model.EstadoModel;
 import com.gerenciador.contas.repository.EstadoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,7 +27,6 @@ public class EstadoService {
     }
 
     public EstadoModel cadastrar(EstadoModel estado){
-        estado.getCodigo();
         return estadoRepository.save(estado);
     }
 
@@ -33,7 +35,13 @@ public class EstadoService {
     }
 
     public List<EstadoModel> deletar(Long id){
-        estadoRepository.deleteById(id);
-        return buscarTodos();
+        if (estadoRepository.findById(id).isPresent()) {
+            estadoRepository.deleteById(id);
+            throw new HttpClientErrorException(HttpStatus.NO_CONTENT);
+        }else if (estadoRepository.findAll().isEmpty()){
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
+        }else
+            throw new HttpServerErrorException(HttpStatus.NOT_FOUND);
     }
+
 }
